+++ webapp/views/App/views/Users/UserEdit/UserExtraPropsEditor/index.js
@@ -1,0 +1,1 @@
+export { UserExtraPropsEditor } from './UserExtraPropsEditor'
